move errorContains to local shrex
